\name{a}
\alias{a}
\docType{data}
\title{dataset a
}
\description{
A concise (1-5 lines) description of the dataset.
}
\usage{data("a")}
\format{
  The format is:
 Time-Series [1:20] from 1977 to 1996: 16094 16368 15234 15371 15684 ...
}
\examples{
data(a)
## maybe str(a) ; plot(a) ...
}
\keyword{datasets}
