\name{ABS}
\alias{ABS}
\docType{data}
\title{dataset ABS
}
\description{
A concise (1-5 lines) description of the dataset. 
}
\usage{data("ABS")}
\format{
  A data frame with 425 observations on the following 22 variables.
  \describe{
    \item{\code{X0.2.09.10.M}}{a ts}
    \item{\code{X0.2.08.10.M}}{a ts}
    \item{\code{X0.2.07.10.M}}{a ts}
    \item{\code{X0.2.06.10.M}}{a ts}
    \item{\code{X0.2.05.10.M}}{a ts}
    \item{\code{X0.2.04.10.M}}{a ts}
    \item{\code{X0.2.03.10.M}}{a ts}
    \item{\code{X0.2.02.10.M}}{a ts}
    \item{\code{X0.2.01.10.M}}{a ts}
    \item{\code{X0.2.15.10.M}}{a ts}
    \item{\code{X0.2.46.10.M}}{a ts}
    \item{\code{X0.2.14.10.M}}{a ts}
    \item{\code{X0.2.45.10.M}}{a ts}
    \item{\code{X0.2.13.10.M}}{a ts}
    \item{\code{X0.2.44.10.M}}{a ts}
    \item{\code{X0.2.12.10.M}}{a ts}
    \item{\code{X0.2.43.10.M}}{a ts}
    \item{\code{X0.2.11.10.M}}{a ts}
    \item{\code{X0.2.42.10.M}}{a ts}
    \item{\code{X0.2.10.10.M}}{a ts}
    \item{\code{X0.2.41.10.M}}{a ts}
    \item{\code{X0.2.20.10.M}}{a ts}
  }
}
\examples{
data(ABS)
## maybe str(ABS) ; plot(ABS) ...
}
\keyword{datasets}
