% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{aggregate}
\alias{aggregate}
\title{Aggregation of time series}
\usage{
aggregate(s, nfreq = 1, conversion = "Sum", complete = TRUE)
}
\arguments{
\item{complete}{}
}
\value{

}
\description{
Aggregation of time series
}
