% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg.R
\docType{methods}
\name{coef,JD3_TemporalDisaggregation-method}
\alias{coef,JD3_TemporalDisaggregation-method}
\title{Coefficients of the regression model}
\usage{
\S4method{coef}{JD3_TemporalDisaggregation}(object)
}
\arguments{
\item{JD3_TemporalDisaggregation}{}
}
\description{
Coefficients of the regression model
}
