\name{jd3_aggregate}
\alias{jd3_aggregate}
\title{jd3_aggregate
}
\description{
A concise (1-5 lines) description of the dataset.
}

\format{
  The format is:
function (s, nfreq = 1, conversion = "Sum", complete = TRUE)  
 - attr(*, "srcref")= 'srcref' int [1:8] 48 16 61 1 16 1 75 88
  ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x0000000019c46b70> 
}
\details{
If necessary, more details than the __description__ above 
}
\source{
reference to a publication or URL from which the data were obtained 
}
\references{
possibly secondary sources and usages
}
