\name{jd3_cholette}
\alias{jd3_cholette}
\title{
cholette benchmarking
}
\description{
A concise (1-5 lines) description of the dataset. 
}

\format{
  The format is:
function (s, t, rho = 1, lambda = 1, bias = "None", conversion = "Sum")  
 - attr(*, "srcref")= 'srcref' int [1:8] 11 15 17 1 15 1 330 336
  ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x0000000019c38540> 
}
\details{
If necessary, more details than the __description__ above
}
\source{
reference to a publication or URL from which the data were obtained
}
\references{
possibly secondary sources and usages 
}
