% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jd3_cubicspline}
\alias{jd3_cubicspline}
\title{Title}
\usage{
jd3_cubicspline(s, t, conversion = "Sum", obsposition = 1)
}
\arguments{
\item{obsposition}{Postion of the observation in the aggregated period (only used with "UserDefined" conversion)}
}
\description{
Title
}
