% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jd3_cubicspline2}
\alias{jd3_cubicspline2}
\title{Benchmarking (without indicator) by means of cubic splines}
\usage{
jd3_cubicspline2(nfreq, t, conversion = "Sum", obsposition = 1)
}
\arguments{
\item{obsposition}{Postion of the observation in the aggregated period (only used with "UserDefined" conversion)}
}
\value{

}
\description{
Benchmarking (without indicator) by means of cubic splines
}
