% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jd3_denton}
\alias{jd3_denton}
\title{Benchmarking by means of the Denton method.}
\usage{
jd3_denton(s, t, d = 1, mul = TRUE, modified = TRUE,
  conversion = "Sum", obsposition = 1)
}
\arguments{
\item{s}{Disaggregated series. Mandatory}

\item{t}{Aggregation constraint. Mandatory}

\item{d}{Differencing order. 1 by default}

\item{mul}{Multiplicative or additive benchmarking. Multiplicative by default}

\item{modified}{Modified (TRUE) or unmodified (FALSE) Denton. Modified by default}

\item{conversion}{Conversion rule. Should be "Sum" or "Average". Sum by default.}

\item{obsposition}{Postion of the observation in the aggregated period (only used with "UserDefined" conversion)}
}
\value{
The benchmarked series is returned
}
\description{
Benchmarking by means of the Denton method.
}
