% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jd3_denton2}
\alias{jd3_denton2}
\title{Benchmarking by means of the Denton method without indicator}
\usage{
jd3_denton2(nfreq, t, d = 1, mul = TRUE, modified = TRUE,
  conversion = "Sum")
}
\arguments{
\item{t}{Aggregation constraint. Mandatory}

\item{d}{Differencing order. 1 by default}

\item{mul}{Multiplicative or additive benchmarking. Multiplicative by default}

\item{modified}{Modified (TRUE) or unmodified (FALSE) Denton. Modified by default}

\item{conversion}{Conversion rule. Should be "Sum" or "Average". Sum by default.}
}
\value{
The benchmarked series is returned
}
\description{
Benchmarking by means of the Denton method without indicator
}
