% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jd3_grp}
\alias{jd3_grp}
\title{Benchmarking following the growth rate preservation principle.
This method corresponds to the method of Cauley and Trager, using the solution
proposed by Di Fonzo and Marini.}
\usage{
jd3_grp(
  s,
  t,
  conversion = "Sum",
  obsposition = 1,
  eps = 1e-12,
  iter = 500,
  denton = T
)
}
\arguments{
\item{obsposition}{Postion of the observation in the aggregated period (only used with "UserDefined" conversion)}
}
\value{

}
\description{
Benchmarking following the growth rate preservation principle.
This method corresponds to the method of Cauley and Trager, using the solution
proposed by Di Fonzo and Marini.
}
