% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jd3_mcholette}
\alias{jd3_mcholette}
\title{Multi-variate Cholette}
\usage{
jd3_mcholette(xlist, tcvector = NULL, ccvector = NULL, rho = 1, lambda = 1)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
Multi-variate Cholette
}
