% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg.R
\name{jd3_tempdisagg}
\alias{jd3_tempdisagg}
\title{Title}
\usage{
jd3_tempdisagg(
  series,
  constant = T,
  trend = F,
  indicators = NULL,
  model = c("Ar1", "Rw", "RwAr1"),
  freq = 4,
  conversion = c("Sum", "Average", "Last", "First", "UserDefined"),
  conversion.obsposition = 1,
  rho = 0,
  rho.fixed = F,
  rho.truncated = 0,
  zeroinitialization = F,
  diffuse.algorithm = c("SqrtDiffuse", "Diffuse", "Augmented"),
  diffuse.regressors = F
)
}
\arguments{
\item{series}{The time series taht will be disaggregated}

\item{constant}{Constant term (T/F)}

\item{trend}{Linear trend (T/F)}

\item{indicators}{High-frequency indicator used in the temporal disaggregation}

\item{model}{Model of the error term (at the disaggregated level). "Ar1" = Chow-Lin, "Rw" = Fernandez, "RwAr1" = Litterman}

\item{freq}{Annual frequency of the disaggregated variable. Used if no indicator is provided}

\item{conversion}{Conversion mode (Usually "Sum" or "Average")}

\item{conversion.obsposition}{Only used with "UserDefined" mode. Position of the observed indicator in the aggregated periods (for instance 7th month of the year)}

\item{rho}{Only used with Ar1/RwAr1 models. Initial value of the parameter}

\item{rho.fixed}{Fixed rho (T/F)}

\item{rho.truncated}{Range for Rho evaluation (in [rho.truncated, 1[)}

\item{zeroinitialization}{Initial values of auto-regressive models are fixed to 0}

\item{diffuse.algorithm}{Algorithm used for diffuse initialization}

\item{diffuse.regressors}{Indicates if the coefficients of the regression model are diffuse (T) or fixed unknown (F, default)}
}
\value{

}
\description{
Temporal disaggregation of a time series
}
