% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg.R
\docType{methods}
\name{logLik,JD3_TemporalDisaggregation-method}
\alias{logLik,JD3_TemporalDisaggregation-method}
\title{Log-likelihood and linked statistics}
\usage{
\S4method{logLik}{JD3_TemporalDisaggregation}(object)
}
\arguments{
\item{JD3_TemporalDisaggregation}{}
}
\description{
Log-likelihood and linked statistics
}
