% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg.R
\docType{methods}
\name{show,JD3_TemporalDisaggregation-method}
\alias{show,JD3_TemporalDisaggregation-method}
\title{Main output of a temporal disaggregation by regression model}
\usage{
\S4method{show}{JD3_TemporalDisaggregation}(object)
}
\arguments{
\item{JD3_TemporalDisaggregation}{}
}
\description{
Main output of a temporal disaggregation by regression model
}
