% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg.R
\docType{methods}
\name{summary,JD3_TemporalDisaggregation-method}
\alias{summary,JD3_TemporalDisaggregation-method}
\title{Detailed output of a temporal disaggregation by regression model}
\usage{
\S4method{summary}{JD3_TemporalDisaggregation}(object)
}
\arguments{
\item{JD3_TemporalDisaggregation}{}
}
\description{
Detailed output of a temporal disaggregation by regression model
}
