% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jd3_cholette}
\alias{jd3_cholette}
\title{Cholette method}
\usage{
jd3_cholette(s, t, rho = 1, lambda = 1, bias = "None",
  conversion = "Sum")
}
\arguments{
\item{s}{Disaggregated series. Mandatory}

\item{t}{Aggregation constraint. Mandatory}

\item{conversion}{}
}
\description{
Benchmarking by means of the Cholette method.
}
\details{
\deqn{\sum_{i,t}\left(\left(\frac{{x_{i,t}-z}_{i,t}}{\left|z_{i,t}\right|^\lambda}\right)-\rho\left(\frac{{x_{i,t-1}-z}_{i,t-1}}{\left|z_{i,t-1}\right|^\lambda}\right)\right)^2}
}
