\name{jd3_tempdisagg}
\alias{jd3_tempdisagg}
\title{
temporal disaggregation
}
\description{
A concise (1-5 lines) description of the dataset. 
}

\format{
  The format is:
function (series, constant = T, trend = F, indicators = NULL, model = "Ar1", freq = 4, 
    conversion = "Sum", conversion.obsposition = 1, rho = 0, rho.fixed = F, rho.truncated = 0, 
    diffuse.algorithm = "SqrtDiffuse", diffuse.zeroinitialization = F, diffuse.regressors = F)  
 - attr(*, "srcref")= 'srcref' int [1:8] 9 17 26 1 17 1 347 364
  ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x0000000019c108f8> 
}
\details{
If necessary, more details than the __description__ above 
}
\source{
reference to a publication or URL from which the data were obtained ~~
}
